refactor: Reorganize Onboarding navigation and hooks
- Move `useToOnBoardingPage` hook to a dedicated file
- Create new GetStarted page with improved structure
- Add TermsAndPrivacy component
- Update import paths for onboarding-related files
- Add 'onboard' to spell checker skip words
