+++ packages/kit/src/components/Empty/EmptyWallet.tsx
@@ -1,7 +1,7 @@
 import { useIntl } from 'react-intl';
 
 import { Empty } from '@onekeyhq/components';
-import { useToOnBoardingPage } from '@onekeyhq/kit/src/views/Onboarding/pages';
+import { useToOnBoardingPage } from '@onekeyhq/kit/src/views/Onboarding/hooks/useToOnBoardingPage';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 
 function EmptyWallet() {
