+++ packages/kit/src/views/Onboarding/hooks/useToOnBoardingPage.ts
@@ -1,0 +1,49 @@
+import { useMemo } from 'react';
+
+import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
+import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
+import platformEnv from '@onekeyhq/shared/src/platformEnv';
+import type { IOnboardingParamList } from '@onekeyhq/shared/src/routes';
+import {
+  EModalRoutes,
+  EOnboardingPages,
+  ERootRoutes,
+} from '@onekeyhq/shared/src/routes';
+
+export const useToOnBoardingPage = () => {
+  const navigation = useAppNavigation();
+
+  return useMemo(
+    () =>
+      async ({
+        isFullModal = false,
+        params,
+      }: {
+        isFullModal?: boolean;
+        params?: IOnboardingParamList[EOnboardingPages.GetStarted];
+      } = {}) => {
+        if (platformEnv.isExtensionUiPopup) {
+          await backgroundApiProxy.serviceApp.openExtensionExpandTab({
+            routes: [
+              isFullModal ? ERootRoutes.iOSFullScreen : ERootRoutes.Modal,
+              EModalRoutes.OnboardingModal,
+              EOnboardingPages.GetStarted,
+            ],
+            params: {
+              ...params,
+              fromExt: true,
+            },
+          });
+        } else {
+          navigation[isFullModal ? 'pushFullModal' : 'pushModal'](
+            EModalRoutes.OnboardingModal,
+            {
+              screen: EOnboardingPages.GetStarted,
+              params,
+            },
+          );
+        }
+      },
+    [navigation],
+  );
+};
