+++ packages/kit/src/views/Onboarding/pages/GetStarted/index.ts
@@ -1,0 +1,1 @@
+export * from './GetStarted';
